Moved some caching stuff around so I can see what depends on what
