+++ nym-api/src/node_status_api/cache/refresher.rs
@@ -1,149 +1,24 @@
-// Copyright 2022 - Nym Technologies SA <contact@nymtech.net>
-// SPDX-License-Identifier: Apache-2.0
+use std::{collections::HashMap, time::Duration};
 
-use crate::contract_cache::{Cache, CacheNotification, ValidatorCache};
-use crate::storage::NymApiStorage;
-use mixnet_contract_common::families::FamilyHead;
-use mixnet_contract_common::reward_params::Performance;
 use mixnet_contract_common::{
-    IdentityKey, Interval, MixId, MixNodeDetails, RewardedSetNodeStatus, RewardingParams,
+    families::FamilyHead, reward_params::Performance, IdentityKey, Interval, MixId, MixNodeDetails,
+    RewardedSetNodeStatus, RewardingParams,
 };
-use nym_api_requests::models::{MixNodeBondAnnotated, MixnodeStatus};
-use rocket::fairing::AdHoc;
-use std::collections::HashMap;
-use std::{sync::Arc, time::Duration};
+use nym_api_requests::models::MixNodeBondAnnotated;
 use task::TaskClient;
-use tokio::sync::RwLockReadGuard;
-use tokio::{
-    sync::{watch, RwLock},
-    time,
+use tokio::sync::watch;
+
+use crate::{
+    contract_cache::cache::{refresher::CacheNotification, ValidatorCache},
+    node_status_api::{
+        cache::{inclusion_probabilities::InclusionProbabilities, NodeStatusCacheError},
+        reward_estimate::{compute_apy_from_reward, compute_reward_estimate},
+    },
+    storage::NymApiStorage,
 };
+use tokio::time;
 
-use self::inclusion_probabilities::InclusionProbabilities;
-
-use super::reward_estimate::{compute_apy_from_reward, compute_reward_estimate};
-
-mod inclusion_probabilities;
-
-const CACHE_TIMEOUT_MS: u64 = 100;
-
-enum NodeStatusCacheError {
-    SimulationFailed,
-    SourceDataMissing,
-}
-
-// A node status cache suitable for caching values computed in one sweep, such as active set
-// inclusion probabilities that are computed for all mixnodes at the same time.
-//
-// The cache can be triggered to update on contract cache changes, and/or periodically on a timer.
-#[derive(Clone)]
-pub struct NodeStatusCache {
-    inner: Arc<RwLock<NodeStatusCacheInner>>,
-}
-
-#[derive(Default)]
-struct NodeStatusCacheInner {
-    mixnodes_annotated: Cache<Vec<MixNodeBondAnnotated>>,
-    rewarded_set_annotated: Cache<Vec<MixNodeBondAnnotated>>,
-    active_set_annotated: Cache<Vec<MixNodeBondAnnotated>>,
-
-    // Estimated active set inclusion probabilities from Monte Carlo simulation
-    inclusion_probabilities: Cache<InclusionProbabilities>,
-}
-
-impl NodeStatusCache {
-    fn new() -> Self {
-        NodeStatusCache {
-            inner: Arc::new(RwLock::new(NodeStatusCacheInner::new())),
-        }
-    }
-
-    pub fn stage() -> AdHoc {
-        AdHoc::on_ignite("Node Status Cache", |rocket| async {
-            rocket.manage(Self::new())
-        })
-    }
-
-    async fn update_cache(
-        &self,
-        mixnodes: Vec<MixNodeBondAnnotated>,
-        rewarded_set: Vec<MixNodeBondAnnotated>,
-        active_set: Vec<MixNodeBondAnnotated>,
-        inclusion_probabilities: InclusionProbabilities,
-    ) {
-        match time::timeout(Duration::from_millis(CACHE_TIMEOUT_MS), self.inner.write()).await {
-            Ok(mut cache) => {
-                cache.mixnodes_annotated.update(mixnodes);
-                cache.rewarded_set_annotated.update(rewarded_set);
-                cache.active_set_annotated.update(active_set);
-                cache
-                    .inclusion_probabilities
-                    .update(inclusion_probabilities);
-            }
-            Err(e) => error!("{e}"),
-        }
-    }
-
-    async fn get_cache<T>(
-        &self,
-        fn_arg: impl FnOnce(RwLockReadGuard<'_, NodeStatusCacheInner>) -> Cache<T>,
-    ) -> Option<Cache<T>> {
-        match time::timeout(Duration::from_millis(CACHE_TIMEOUT_MS), self.inner.read()).await {
-            Ok(cache) => Some(fn_arg(cache)),
-            Err(e) => {
-                error!("{e}");
-                None
-            }
-        }
-    }
-
-    pub(crate) async fn mixnodes_annotated(&self) -> Option<Cache<Vec<MixNodeBondAnnotated>>> {
-        self.get_cache(|c| c.mixnodes_annotated.clone()).await
-    }
-
-    pub(crate) async fn rewarded_set_annotated(&self) -> Option<Cache<Vec<MixNodeBondAnnotated>>> {
-        self.get_cache(|c| c.rewarded_set_annotated.clone()).await
-    }
-
-    pub(crate) async fn active_set_annotated(&self) -> Option<Cache<Vec<MixNodeBondAnnotated>>> {
-        self.get_cache(|c| c.active_set_annotated.clone()).await
-    }
-
-    pub(crate) async fn inclusion_probabilities(&self) -> Option<Cache<InclusionProbabilities>> {
-        self.get_cache(|c| c.inclusion_probabilities.clone()).await
-    }
-
-    pub async fn mixnode_details(
-        &self,
-        mix_id: MixId,
-    ) -> (Option<MixNodeBondAnnotated>, MixnodeStatus) {
-        // it might not be the most optimal to possibly iterate the entire vector to find (or not)
-        // the relevant value. However, the vectors are relatively small (< 10_000 elements, < 1000 for active set)
-
-        let active_set = &self.active_set_annotated().await.unwrap().into_inner();
-        if let Some(bond) = active_set.iter().find(|mix| mix.mix_id() == mix_id) {
-            return (Some(bond.clone()), MixnodeStatus::Active);
-        }
-
-        let rewarded_set = &self.rewarded_set_annotated().await.unwrap().into_inner();
-        if let Some(bond) = rewarded_set.iter().find(|mix| mix.mix_id() == mix_id) {
-            return (Some(bond.clone()), MixnodeStatus::Standby);
-        }
-
-        let all_bonded = &self.mixnodes_annotated().await.unwrap().into_inner();
-        if let Some(bond) = all_bonded.iter().find(|mix| mix.mix_id() == mix_id) {
-            (Some(bond.clone()), MixnodeStatus::Inactive)
-        } else {
-            (None, MixnodeStatus::NotFound)
-        }
-    }
-}
-
-impl NodeStatusCacheInner {
-    pub fn new() -> Self {
-        Self::default()
-    }
-}
+use super::NodeStatusCache;
 
 // Long running task responsible of keeping the cache up-to-date.
 pub struct NodeStatusCacheRefresher {
