+++ nym-api/src/node_status_api/mod.rs
@@ -1,7 +1,7 @@
 // Copyright 2021 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: Apache-2.0
 
-pub(crate) use cache::{NodeStatusCache, NodeStatusCacheRefresher};
+pub(crate) use cache::NodeStatusCache;
 use okapi::openapi3::OpenApi;
 use rocket::Route;
 use rocket_okapi::{openapi_get_routes_spec, settings::OpenApiSettings};
